requirements.txt
discord.py~=2.3
paramiko~=3.5
python-dotenv~=1.0
psutil~=6.1
fastapi~=0.109.0
uvicorn[standard]~=0.27.0
jinja2~=3.1.3
python-multipart~=0.0.9
cryptography~=42.0.2
itsdangerous~=2.1.2
httpx~=0.27.0
